fix delete

+++ app/Http/Controllers/Owner/BanksController.php
@@ -116,7 +116,8 @@ class BanksController extends Controller
             $bankLogo['path'] = $pathLogo;
             $deleteOldLogo = $bank->path;
 
-            if(file_exists(public_path($deleteOldLogo))) {
+
+            if(file_exists(public_path($deleteOldLogo)) && $deleteOldLogo != "/images/banks_logo/") {
                 unlink(public_path($deleteOldLogo));
             }
             $bank->update($bankLogo);
@@ -235,7 +236,7 @@ class BanksController extends Controller
 
         $delete = BankLogo::where('id', $logId);
 
-        if(file_exists(public_path($delete->first()->path))){
+        if(file_exists(public_path($delete->first()->path)) &&  $delete->first()->path != "/images/banks_logo/"  ){
             unlink(public_path($delete->first()->path));
             $delete->delete();
         }else{
